move sidebar to component
